Move data objects in to their own files
