+++ Frodo/Service/DataSyncService.cs
@@ -1,5 +1,6 @@
 ﻿using Gluten.Core.DataProcessing.Service;
 using Gluten.Core.Service;
+using Gluten.Data;
 using Gluten.Data.ClientModel;
 using Gluten.Data.TopicModel;
