+++ Gluten.Core/Service/PinHelper.cs
@@ -1,4 +1,4 @@
-﻿using Gluten.Data.TopicModel;
+﻿using Gluten.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
