+++ Gluten.Data/Gluten.Data.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="PinCache\" />
+  </ItemGroup>
+
 </Project>
