+++ Gluten.Data/PinCacheMeta.cs
@@ -1,0 +1,19 @@
+﻿namespace Gluten.Data
+{
+    public class PinCacheMeta
+    {
+        public string? RestaurantType { get; set; }
+        public string? Price { get; set; }
+        public string? Stars { get; set; }
+
+        public bool PermanentlyClosed { get; set; }
+        // Opening times 
+        // review summary
+        // website 
+        // opening times 
+        // phone no
+        // address 
+        // Dine-in, Takeaway, Delivery?
+
+    }
+}
